correction du contrôler pour verrouiller impossibilité de s'autosupprimer

+++ src/Controller/UserController.php
@@ -90,12 +90,19 @@ class UserController extends AbstractController
      */
     public function delete(Request $request, User $user): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($user);
-            $em->flush();
+        if ($this->getUser() === $user) {
+            $this->addFlash(
+                'notice',
+                'Cette action est impossible !'
+            );
+        } else {
+            if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($user);
+                $em->flush();
+            }
+
+            return $this->redirectToRoute('user_index');
         }
-        
-        return $this->redirectToRoute('user_index');
     }
 }
